Move debug stuff to debug folder
